requirements.txt
openai
groq
requests
google-generativeai
cohere
python-dotenv